+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Documentation for TransferBench is available at
 - Reporting number of iterations run when running in timed mode (NUM_ITERATIONS < 0)
 - Adding NIC_CQ_POLL_BATCH to control CQ poll batch size for NIC transfers
 - New "hbm" preset which sweeps and tests local HBM read performance
+- Added a new TB_WALLCLOCK_RATE that will override GPU GFX wallclock rate if it returns 0 (debug)
 
 ### Modified
   - DMA-BUF support enablement in CMake changed to ENABLE_DMA_BUF to be more similar to other compile-time options